refactored i2c driver
