+++ DRIVERS/MSP430/UTILS/utils.c
@@ -1,0 +1,13 @@
+
+
+#include "utils.h"
+
+// Helper function to copy buffer contents to specific command buffer
+void CopyArray(uint8_t *source, uint8_t *dest, uint8_t count)
+{
+    uint8_t copyIndex = 0;
+    for (copyIndex = 0; copyIndex < count; copyIndex++)
+    {
+        dest[copyIndex] = source[copyIndex];
+    }
+}
