+++ DRIVERS/MSP430/UTILS/utils.h
@@ -1,0 +1,11 @@
+
+#include <stdint.h>
+
+/**
+ * @brief Copies a specified number of bytes from a source array to a destination array.
+ *
+ * @param source Pointer to the source array
+ * @param dest Pointer to the destination array
+ * @param count Number of bytes to copy
+ */
+void CopyArray(uint8_t *source, uint8_t *dest, uint8_t count);
